requirements.txt
google-ai-generativelanguage==0.4.0
google-api-core==2.12.0
google-auth==2.23.3
google-generativeai==0.4.1
pyppeteer==1.0.2